content/canvas/test/webgl/non-conf-tests/test_highp_fs.html

<!DOCTYPE HTML>
<title>WebGL test: `highp` support</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<script src="webgl-util.js"></script>
<script src="driver-info.js"></script>
<script id="shader-vs" type="x-shader/x-vertex">

void main(void) {
  gl_Position = vec4(vec3(0.0), 1.0);
}

</script>
<script id="shader-fs" type="x-shader/x-fragment">

precision highp float;

void main(void) {
  gl_FragColor = vec4(1.0, 1.0, 1.0, 1.0);
}

</script>
<body>
<canvas id="c"></canvas>
<script>

var gl = WebGLUtil.getWebGL('c');

// Give ourselves a scope to return early from:
(function() {
  if (DriverInfo.getOS() == DriverInfo.OS.ANDROID &&
      DriverInfo.getOSVersion() < 15)
  {
    // Consider 'random'. Actually, ARMv6 fails, and ARMv7 succeeds, but we have
    // not been successful at determining this from JS. (see bug 917478)
    return;
  }

  ok(gl, 'Should have a WebGL context.');

  // Catch actual WebGLUtil errors, not GL errors.
  function errorFunc(str) {
    ok(false, 'Error: ' + str);
  }
  WebGLUtil.setErrorFunc(errorFunc);

  var format = gl.getShaderPrecisionFormat(gl.FRAGMENT_SHADER, gl.HIGH_FLOAT);
  ok(!gl.getError(), 'Unexpected GL error.');
  var prog = WebGLUtil.createProgramByIds(gl, 'shader-vs', 'shader-fs');
  ok(!gl.getError(), 'Unexpected GL error.');
  
  if (format) {
    ok(prog, 'Frag shader with unconditional `precision highp float` should ' +
             'link if `getShaderPrecisionFormat` gives a format for it.');
  } else {
    ok(!prog, 'Frag shader with unconditional `precision highp float` should ' +
              'NOT link if `getShaderPrecisionFormat` gives NO format for it.');
  }
})();

</script>
